Create routines to save a state file for Paraview.

+++ fileops/export/_bioformats.py
@@ -1,35 +1,48 @@
 import os.path
-from typing import List
+from pathlib import Path
+from typing import List, Dict
 
 import numpy as np
 from tifffile import imwrite
 
-from fileops.cached import CachedImageFile
 from fileops.cached.cached_image_file import ensure_dir
-from fileops.image import OMEImageFile
+from fileops.image import ImageFile, OMEImageFile
 from fileops.image.exceptions import FrameNotFoundError
 from fileops.logger import get_logger
 
 log = get_logger(name='export')
 
 
-def bioformats_to_tiffseries(path, img_struct: CachedImageFile, save_folder='_vol_paraview'):
+def bioformats_to_tiffseries(img_struct: ImageFile, save_path=Path('_vol_paraview')) -> Dict:
     log.info("Exporting bioformats file to series of tiff file volumes.")
-    base_dir = os.path.dirname(path)
-    sav_path = os.path.join(base_dir, save_folder)
-    ensure_dir(os.path.join(sav_path, 'dummy'))
+    save_path = ensure_dir(save_path)
 
+    out = dict()
     image = np.empty(shape=(len(img_struct.zstacks), *img_struct.image(0).image.shape), dtype=np.uint16)
     for j, c in enumerate(img_struct.channels):
+        out[f"ch{c:01d}"] = {
+            "files":  [],
+            "minmax": []
+        }
+        ensure_dir(save_path / f"ch{c:01d}")
         for fr in img_struct.frames:
-            for i, z in enumerate(img_struct.zstacks):
-                ix = img_struct.ix_at(c=j, z=z, t=fr)
-                mdimg = img_struct.image(ix)
-                image[i, :, :] = mdimg.image
             fname = f'C{c:02d}T{fr:04d}_vol.tiff'
-            fpath = os.path.join(sav_path, fname)
+            fpath = save_path / f"ch{c:01d}" / fname
+
+            out[f"ch{c:01d}"]["files"].append(fpath.as_posix())
+            out[f"ch{c:01d}"]["minmax"].append((np.min(image), np.max(image)))
+
             log.debug(f"Saving image {fname} in cache (path={fpath}).")
-            imwrite(fpath, np.array(image), imagej=True, metadata={'order': 'ZXY'})
+            if not os.path.exists(fpath):
+                for i, z in enumerate(img_struct.zstacks):
+                    ix = img_struct.ix_at(c=j, z=z, t=fr)
+                    mdimg = img_struct.image(ix)
+                    image[i, :, :] = mdimg.image
+                imwrite(fpath, np.array(image), imagej=True, metadata={'order': 'ZXY'})
+            else:
+                print(f"skipping file {fpath.as_posix()}")
+
+    return out
 
 
 def bioformats_to_ndarray_zstack(img_struct: OMEImageFile, roi=None, channel=0, frame=0):
